iOS8 searchbar fix + minor improvements

+++ Q-municate/Scenes/Chat/QMChatVC/QMChatVC.m
@@ -555,7 +555,7 @@ QMImageViewDelegate
         CGSize bottomLabelSize = [TTTAttributedLabel sizeThatFitsAttributedString:attributedString
                                                                   withConstraints:CGSizeMake(MIN(kQMAttachmentCellSize, maxWidth), CGFLOAT_MAX)
                                                            limitedToNumberOfLines:0];
-        size = CGSizeMake(MIN(kQMAttachmentCellSize, maxWidth), kQMAttachmentCellSize + ceil(bottomLabelSize.height));
+        size = CGSizeMake(MIN(kQMAttachmentCellSize, maxWidth), kQMAttachmentCellSize + ceilf(bottomLabelSize.height));
     }
     else if (viewClass == [QMChatNotificationCell class]) {
